Emily's first updates
Hi Yigit- Great and informative tutorial! I made mostly grammatical changes, but there are a few things to note. I left inline comments using HTML but if you search "<!--EM" you'll be able to see all of my questions and comments. The GIFs are great and super informative but for accessibility reasons and SEO reasons, we need to write out the steps that are demonstrated in the GIFs. Let me know if you have any questions about that!